+++ app/src/main/java/com/tyroo/tyroovastdemoapp/CarousalActivity.java
@@ -5,6 +5,7 @@ import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 
 import com.tyroo.tva.sdk.AdView;
+import com.tyroo.tva.sdk.ErrorCode;
 import com.tyroo.tva.sdk.TyrooVidAiSdk;
 
 public class CarousalActivity extends AppCompatActivity implements TyrooVidAiSdk.TyrooAdListener{
@@ -26,7 +27,7 @@ public class CarousalActivity extends AppCompatActivity implements TyrooVidAiSdk
 
     private void initTyrooVidAiSdk() {
         try {
-            tyrooVidAiSdk = TyrooVidAiSdk.initialize(getApplicationContext(), "1635","009",this);
+            tyrooVidAiSdk = TyrooVidAiSdk.initialize(this, "1635", "009", this);
             tyrooVidAiSdk.setAdViewLayout(adView);
             tyrooVidAiSdk.enableCaching(true);
             tyrooVidAiSdk.loadAds();
@@ -39,17 +40,19 @@ public class CarousalActivity extends AppCompatActivity implements TyrooVidAiSdk
 
     @Override
     protected void onDestroy() {
-        //InitiateTyrooSdk.destroyInstance();
+        super.onDestroy();
         // RefWatcher refWatcher = DemoApplication.getRefWatcher(this);
         // refWatcher.watch(this);
         adView.removeAllViews();
         tyrooVidAiSdk.flush();
-        super.onDestroy();
     }
 
     @Override
     public void onAdLoaded(String placementId) {
-        Log.d(TAG, "onAdLoaded: "+placementId);
+        Log.d(TAG, "onAdLoaded: " + placementId);
+        if (tyrooVidAiSdk.isAdLoaded()) {
+            tyrooVidAiSdk.showAds();
+        }
     }
 
     @Override
@@ -83,8 +86,23 @@ public class CarousalActivity extends AppCompatActivity implements TyrooVidAiSdk
     }
 
     @Override
-    public void onFailure(String errorMsg) {
+    public void onFailure(int errorCode, String errorMsg) {
         Log.e(TAG, "onFailure: " + errorMsg);
+        switch (errorCode){
+            case ErrorCode.BAD_REQUEST:
+                //It may be due to some invalid/Blank value in the api request.
+                break;
+            case ErrorCode.NETWORK_ERROR:
+                //It may be due to slow/no internet connectivity.
+                break;
+            case ErrorCode.NO_INVENTORY:
+                //There is no fill.
+                break;
+            case ErrorCode.UNKNOWN:
+                //In case of any unexpected error or exception.
+                break;
+                default:
+
+        }
     }
-    
 }