+++ app/src/main/java/com/tyroo/tyroovastdemoapp/RecyclerViewAdapter.java
@@ -11,6 +11,7 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.tyroo.tva.sdk.AdView;
+import com.tyroo.tva.sdk.ErrorCode;
 import com.tyroo.tva.sdk.TyrooVidAiSdk;
 import com.tyroo.tva.utils.TyrooLog;
 
@@ -70,10 +71,13 @@ public class RecyclerViewAdapter extends RecyclerView.Adapter<RecyclerView.ViewH
             }
         }
 
-        
+
         @Override
         public void onAdLoaded(String placementId) {
-            Log.d(TAG, "onAdLoaded: "+placementId);
+            Log.d(TAG, "onAdLoaded: " + placementId);
+            if (tyrooVidAiSdk.isAdLoaded()) {
+                tyrooVidAiSdk.showAds();
+            }
         }
 
         @Override
@@ -107,8 +111,24 @@ public class RecyclerViewAdapter extends RecyclerView.Adapter<RecyclerView.ViewH
         }
 
         @Override
-        public void onFailure(String errorMsg) {
+        public void onFailure(int errorCode, String errorMsg) {
             Log.e(TAG, "onFailure: " + errorMsg);
+            switch (errorCode){
+                case ErrorCode.BAD_REQUEST:
+                    //It may be due to some invalid/Blank value in the api request.
+                    break;
+                case ErrorCode.NETWORK_ERROR:
+                    //It may be due to slow/no internet connectivity.
+                    break;
+                case ErrorCode.NO_INVENTORY:
+                    //There is no fill.
+                    break;
+                case ErrorCode.UNKNOWN:
+                    //In case of any unexpected error or exception.
+                    break;
+                default:
+
+            }
         }
 
     }